Add a link to the editorial web search, on the material search page. KBHBIB-76
When on the material search, the default DPL setup creates a "Look for
this query on the website" link, if we find any results.
If the `main` CludoProfile is enabled, we want the link to instead lead
here, along with the query.

To begin with, the link is always shown, without any context if any
results are available.
In the future, once Cludo API has been implemented, we will add this.