Introduce simple/advanced SERPs for Cludo search pages.
Most of the Cludo searches have filters, that need to be sent along as
part of the template.
We'll add an option that can be set dynamically, along with a template.

Ideally, we would want to do the alternative template as a theme hook
suggestion, but apparantly this logic doesnt work in a standalone
modules template folder.