Implement advanced script loading for Cludo.
The Cludo scripts needs to be loaded in a very specific order.
Originally, the hope was that it could just be done by adding them in
the right order in the DOM, but apparantly this was not enough;
Instead, we'll implement a js `Promise` solution, that uses `await`
before loading the next script.